backend/requirements.txt
# AI/ML Services
openai==1.61.1
anthropic==0.42.0

# Web Framework
fastapi==0.115.6
fastapi-cli==0.0.6
uvicorn==0.34.0
httptools==0.6.4
python-multipart==0.0.19

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic_core==2.27.1

# Rate Limiting & Security
slowapi==0.1.9
limits==3.14.1
PyJWT==2.10.1

# HTTP & API
aiohttp==3.11.12
httpx==0.28.1
requests==2.32.3

# Token Counting
tiktoken==0.8.0


# Analytics
api-analytics==1.2.5

# Other dependencies
click==8.1.7
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
PyYAML==6.0.2
rich==13.9.4
tqdm==4.67.1
typer==0.15.1
certifi==2024.12.14
